longevity/requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
celery==5.3.5
certifi==2024.2.2
django-cors-headers==3.10.0
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
django-rest-swagger==2.2.0
django_debug_toolbar==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.7
flake8==7.0.0
gunicorn==20.1.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.6
MarkupSafe==2.1.5
mccabe==0.7.0
openapi-codec==1.3.2
packaging==24.0
pluggy==1.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytils==0.4.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
